chore: remove redundant abstractions
