+++ src/core/RpcSchema.ts
@@ -12,7 +12,6 @@ export type Schema =
       {
         Request: {
           method:
-            | 'wallet_addFaucetFunds'
             | 'wallet_getCapabilities'
             | 'wallet_getAssets'
             | 'wallet_getCallsStatus'
@@ -27,10 +26,6 @@ export type Schema =
       }
     >
   | RpcSchema.From<
-      | {
-          Request: typeof Rpc.wallet_addFaucetFunds.Request.Encoded
-          ReturnType: typeof Rpc.wallet_addFaucetFunds.Response.Encoded
-        }
       | {
           Request: typeof Rpc.account_verifyEmail.Request.Encoded
           ReturnType: typeof Rpc.account_verifyEmail.Response.Encoded
