+++ src/core/internal/mode.ts
@@ -49,13 +49,6 @@ export type Mode = {
       value?: string | undefined
     }) => Promise<{ id: Hex.Hex }>
 
-    addFaucetFunds: (
-      parameters: RpcSchema.wallet_addFaucetFunds.Parameters & {
-        /** Internal properties. */
-        internal: ActionsInternal
-      },
-    ) => Promise<RpcSchema.wallet_addFaucetFunds.Response>
-
     createAccount: (parameters: {
       /** Admins to grant. */
       admins?: readonly Pick<Key.Key, 'publicKey' | 'type'>[] | undefined
