+++ src/core/internal/schema/request.ts
@@ -10,7 +10,6 @@ export * from './rpc.js'
 
 export const Request = Schema.Union(
   RpcRequest.account_verifyEmail.Request,
-  RpcRequest.wallet_addFaucetFunds.Request,
   RpcRequest.wallet_addFunds.Request,
   RpcRequest.eth_accounts.Request,
   RpcRequest.eth_chainId.Request,
